fix(rs): address PR #114 review feedback
- core::time: new shared module — `parse_iso8601_secs`,
  `iso8601_from_unix_secs`, `unix_secs_to_civil`, `days_since_epoch`,
  `now_iso8601`. Single source of truth for the narrow ISO-8601 subset
  the project writes; both `core::session` and `core::claude_telemetry`
  now consume it. Removes the duplicated parsing/epoch helpers.
- Stricter parser: rejects out-of-range hour/minute/second (sec == 60
  still allowed for leap-second wallclock readings), validates
  month-aware day-of-month with proper Gregorian leap-year rules,
  rejects extra `:` / `-` segments, and forces the fractional
  component to ASCII digits only so `NaN` / `+inf` / scientific
  notation can no longer poison `partial_cmp`-based sorts.
- Deterministic retention: `core::session::SessionManager::apply_retention`
  now buckets sessions in a `BTreeMap<String, Vec<Session>>` keyed by
  `worktree_id`, so iteration order — and therefore the persisted
  `projects.json` ordering after a sweep — is stable across runs.
- Doc fix: the `iso8601_from_unix_secs` doc no longer claims byte
  parity with C# `DateTimeOffset.UtcNow.ToString("o")`; it spells out
  that we emit the `Z` shorthand instead of an explicit `+00:00`
  offset and both round-trip through `parse_iso8601_secs`.
