+++ codescope-rs/core/src/lib.rs
@@ -22,6 +22,7 @@ pub mod projects;
 pub mod session;
 pub mod settings;
 pub mod theme;
+pub mod time;
 pub mod window_state;
 
 pub use claude_discovery::{AdoptionCandidate, POLL_INTERVAL_MS as CLAUDE_DISCOVERY_POLL_MS};
